Spectra acquired with Bentham spectrometer

A collection of spectra measured with a Bentham spectroradiometer. Datasets contain wavelengths at a 0.5 nm interval and corresponding values for spectral emittance from different lamps. As absolute values are not meaningful because measuring distances are in most cases unknown data have been normalized to one at the wavelength of maximum energy irradiance.

bentham

Format

A "source.mspct" with two member "source.spct objects", each with 301 rows (250 nm to 400 nm, 0.5 nm step) and 2 variables.

Details

licor is a character vector with indexes to members of lamps.mspct philips.tl01 philips.tl12

The variables in each member spectrum are as follows:

w.length (nm)

s.e.irrad (relative energy based units)

Note

Instrument used: Bentham scanning double monochromator spectroradiometer with a cosine corrected input optics. Recently calibrated at STUK (Radiation Safety Authority Finland). Measurements done by Lasse Ylianttila (STUK, Helsinki, Finland).

Examples

lamps.mspct[bentham]

#> Object: source_mspct [2 x 1]
#> --- Member: philips.tl01 ---
#> Object: source_spct [301 x 2]
#> Wavelength range 250 to 400 nm, step 0.5 nm 
#> Time unit 1s
#> Spectral data normalized to 1 at 311.5 nm 
#> 
#> # A tibble: 301 x 2
#>    w.length    s.e.irrad
#>       <dbl>        <dbl>
#>  1      250 0.000000249 
#>  2      250 0.0000000624
#>  3      251 0.0000000340
#>  4      252 0.0000000166
#>  5      252 0.0000000261
#>  6      252 0.000000468 
#>  7      253 0.0000280   
#>  8      254 0.0000979   
#>  9      254 0.0000102   
#> 10      254 0.000000201 
#> # ... with 291 more rows
#> --- Member: philips.tl12 ---
#> Object: source_spct [301 x 2]
#> Wavelength range 250 to 400 nm, step 0.5 nm 
#> Time unit 1s
#> Spectral data normalized to 1 at 313 nm 
#> 
#> # A tibble: 301 x 2
#>    w.length   s.e.irrad
#>       <dbl>       <dbl>
#>  1      250 0.000000683
#>  2      250 0          
#>  3      251 0          
#>  4      252 0          
#>  5      252 0.000000365
#>  6      252 0.00000135 
#>  7      253 0.0000379  
#>  8      254 0.000153   
#>  9      254 0.0000109  
#> 10      254 0.000000978
#> # ... with 291 more rows
#> 
#> --- END ---

lamps.mspct[bentham]

#> Object: source_mspct [2 x 1]
#> --- Member: philips.tl01 ---
#> Object: source_spct [301 x 2]
#> Wavelength range 250 to 400 nm, step 0.5 nm 
#> Time unit 1s
#> Spectral data normalized to 1 at 311.5 nm 
#> 
#> # A tibble: 301 x 2
#>    w.length    s.e.irrad
#>       <dbl>        <dbl>
#>  1      250 0.000000249 
#>  2      250 0.0000000624
#>  3      251 0.0000000340
#>  4      252 0.0000000166
#>  5      252 0.0000000261
#>  6      252 0.000000468 
#>  7      253 0.0000280   
#>  8      254 0.0000979   
#>  9      254 0.0000102   
#> 10      254 0.000000201 
#> # ... with 291 more rows
#> --- Member: philips.tl12 ---
#> Object: source_spct [301 x 2]
#> Wavelength range 250 to 400 nm, step 0.5 nm 
#> Time unit 1s
#> Spectral data normalized to 1 at 313 nm 
#> 
#> # A tibble: 301 x 2
#>    w.length   s.e.irrad
#>       <dbl>       <dbl>
#>  1      250 0.000000683
#>  2      250 0          
#>  3      251 0          
#>  4      252 0          
#>  5      252 0.000000365
#>  6      252 0.00000135 
#>  7      253 0.0000379  
#>  8      254 0.000153   
#>  9      254 0.0000109  
#> 10      254 0.000000978
#> # ... with 291 more rows
#> 
#> --- END ---

lamps.mspct[["philips.tl01"]]

#> Object: source_spct [301 x 2]
#> Wavelength range 250 to 400 nm, step 0.5 nm 
#> Time unit 1s
#> Spectral data normalized to 1 at 311.5 nm 
#> 
#> # A tibble: 301 x 2
#>    w.length    s.e.irrad
#>       <dbl>        <dbl>
#>  1      250 0.000000249 
#>  2      250 0.0000000624
#>  3      251 0.0000000340
#>  4      252 0.0000000166
#>  5      252 0.0000000261
#>  6      252 0.000000468 
#>  7      253 0.0000280   
#>  8      254 0.0000979   
#>  9      254 0.0000102   
#> 10      254 0.000000201 
#> # ... with 291 more rows

lamps.mspct$philips.tl01

#> Object: source_spct [301 x 2]
#> Wavelength range 250 to 400 nm, step 0.5 nm 
#> Time unit 1s
#> Spectral data normalized to 1 at 311.5 nm 
#> 
#> # A tibble: 301 x 2
#>    w.length    s.e.irrad
#>       <dbl>        <dbl>
#>  1      250 0.000000249 
#>  2      250 0.0000000624
#>  3      251 0.0000000340
#>  4      252 0.0000000166
#>  5      252 0.0000000261
#>  6      252 0.000000468 
#>  7      253 0.0000280   
#>  8      254 0.0000979   
#>  9      254 0.0000102   
#> 10      254 0.000000201 
#> # ... with 291 more rows