Spectral data for Lamps of different types

Names of datasets containing the wavelengths and tabulated values spectral emittance for the lamps of different types.

incandescent_lamps

led_lamps

mercury_lamps

multimetal_lamps

sodium_lamps

Format

A vector of character strings.

Note

In the case of LED-based lamps we include here only ready built commercially available lamps. In some cases assembled from multiple descrete LEDs, possible of mixed types and spectral output. For emission spectra for LEDs available as electronic components please see photobiologyLEDs-package.

See also

lamps.mspct

Examples

sodium_lamps

#> [1] "osram.super.vialox"

multimetal_lamps

#> [1] "osram.hqit.400w"

mercury_lamps

#>  [1] "airam.cf.15w.2700k"    "airam.spiraali"        "germicidal"           
#>  [4] "osram.36w.25"          "osram.l36w.840"        "philips.pls11w.827"   
#>  [7] "philips.tl01"          "philips.tl12"          "philips.tl12.mc"      
#> [10] "philips.tl5.35w.830he" "philips.tld.36w.18"    "philips.tld36w.15"    
#> [13] "philips.tld36w.18"     "philips.tld36w.83"     "philips.tld36w.865"   
#> [16] "philips.tld36w.89"     "philips.tld36w.92"     "philips.tll36w.950"   
#> [19] "qpanel.uva340"         "qpanel.uvb313"         "sylvania.215w.vho"    

led_lamps

#> [1] "airam.led.11w.4000k"             "airam.led.oiva.9w.3000k"        
#> [3] "amaran.al.m9"                    "osram.led.10w.2700k.classicstar"
#> [5] "osram.led.8w.2700k"              "toshiba.led.9.5w.2700k"         
#> [7] "v.light.led.2w.6000k"            "valoya.b100.ap67"               

# select lamps emitting in the amber, yellow, orange region
lamps.mspct[sodium_lamps]

#> Object: source_mspct [1 x 1]
#> --- Member: osram.super.vialox ---
#> Object: source_spct [301 x 2]
#> Wavelength range 300 to 900 nm, step 2 nm 
#> Label: File: Osram.Super.Vialox.PRN 
#> Measured on 0000-10-17 13:54:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 820 nm 
#> 
#> # A tibble: 301 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1     300.  0.000235
#>  2     302.  0.000196
#>  3     304.  0.000237
#>  4     306.  0.000378
#>  5     308.  0.000553
#>  6     310.  0.000902
#>  7     312.  0.00139 
#>  8     314.  0.00198 
#>  9     316.  0.00270 
#> 10     318.  0.00348 
#> # ... with 291 more rows
#> 
#> --- END ---