Spectral irradiance for diverse lamps

A collection of lamp emission spectra measured with different spectroradiometers.

lamps.mspct

Format

A "source_mspct" object containing several "source_spct".

Details

The "source_mspct" object contains "source_spct" objects with spectral emission data.

The variables in each member spectrum are as follows:

w.length (nm)

s.e.irrad (relative energy based units)

Note

Please see the help corresponding to each instrument for details.

See also

bentham_lamps, licor_lamps, macam_lamps, oo_maya_lamps for information on the spectroradiometers, and photobiologyLEDs for spectra for LEDs available as electronic components.

Examples

lamps.mspct$qpanel.uvb313

#> Object: source_spct [561 x 2]
#> Wavelength range 240 to 800 nm, step 1 nm 
#> Time unit 1s
#> Spectral data normalized to 1 at 435 nm 
#> 
#> # A tibble: 561 x 2
#>    w.length s.e.irrad
#>       <int>     <dbl>
#>  1      240         0
#>  2      241         0
#>  3      242         0
#>  4      243         0
#>  5      244         0
#>  6      245         0
#>  7      246         0
#>  8      247         0
#>  9      248         0
#> 10      249         0
#> # ... with 551 more rows