Spectra for a multichannel LED bulb

A collection of lamp emission spectra for a household LED bulb with four channels and a remote control allowing color mixing. In addition character vectors useful as indexes for extracting subsets of the data are defined. Pairs of RGB channels are used in mixes. This is by design of the lamp and/or its remote control. The lamp here used is a 7.6 W, 230 V, E-27 LED lamp bulb accompanied by a small wireless remote control from a Swedish supplier. Similar (identical?) lamps are at the moment (early 2018) widely available through eBay and Aliexpress sellers.

ledsavers.mspct

ledsavers_channels

ledsavers_mixes

ledsavers_RG_mixes

ledsavers_RB_mixes

ledsavers_GB_mixes

Format

ledsavers.mspct is a "source_mspct" object containing a collection of 16 "source_spct" objects as members. Members are named. ledsavers_channels, ledsavers_mixes, ledsavers_RG_mixes, ledsavers_RB_mixes, ledsavers_GB_mixes are character vectors containing subsets of names(ledsavers.mspct).

Details

ledsavers.mspct contains a collection of 16 "source_spct" objects with spectral emission data for each of the colors obtainable from the lamp. All colors were measured at maximum power output although the lamp supports dimming. The lamp was maintained at the same distance from the input optics so the different spectra are comparable and provided expressed in calibrated units.

The variables in each member spectrum are as follows:

w.length (nm)

s.e.irrad (relative energy based units)

Note

Instrument used: Ocean Optics Maya2000 Pro single-monochromator array spectroradiometer with a Bentham cosine corrected input optics. A complex set of corrections and calibration procedure used. Raw spectral data processed with R package MayaCalc. The source_spct objects have attributes with additional information on the measurement and data processing. Measurements done by Pedro J. Aphalo. Data acquired using R package 'ooacquire'.

Acknowledgements

We thank Prof. Lars Olof Björn for the gift of the lamp.

References

http://www.ledsavers.com/

Examples

ledsavers_channels

#> [1] "W" "R" "G" "B"

ledsavers.mspct$W

#> Object: source_spct [1,212 x 2]
#> Wavelength range 350 to 898.81 nm, step 0.36 to 0.47 nm 
#> Label: LedSaver_7.5W_W 
#> Measured on 2017-10-02 12:39:51 UTC 
#> Time unit 1s
#> 
#> # A tibble: 1,212 x 2
#>    w.length s.e.irrad
#>  *    <dbl>     <dbl>
#>  1      350  0.000620
#>  2      350  0.000536
#>  3      351  0.000627
#>  4      351  0.000566
#>  5      352  0.000553
#>  6      352  0.000461
#>  7      353  0.000554
#>  8      353  0.000641
#>  9      354  0.000595
#> 10      354  0.000457
#> # ... with 1,202 more rows

ledsavers.mspct[["W"]]

#> Object: source_spct [1,212 x 2]
#> Wavelength range 350 to 898.81 nm, step 0.36 to 0.47 nm 
#> Label: LedSaver_7.5W_W 
#> Measured on 2017-10-02 12:39:51 UTC 
#> Time unit 1s
#> 
#> # A tibble: 1,212 x 2
#>    w.length s.e.irrad
#>  *    <dbl>     <dbl>
#>  1      350  0.000620
#>  2      350  0.000536
#>  3      351  0.000627
#>  4      351  0.000566
#>  5      352  0.000553
#>  6      352  0.000461
#>  7      353  0.000554
#>  8      353  0.000641
#>  9      354  0.000595
#> 10      354  0.000457
#> # ... with 1,202 more rows

ledsavers.mspct[ledsavers_channels]

#> Object: source_mspct [4 x 1]
#> --- Member: W ---
#> Object: source_spct [1,212 x 2]
#> Wavelength range 350 to 898.81 nm, step 0.36 to 0.47 nm 
#> Label: LedSaver_7.5W_W 
#> Measured on 2017-10-02 12:39:51 UTC 
#> Time unit 1s
#> 
#> # A tibble: 1,212 x 2
#>    w.length s.e.irrad
#>  *    <dbl>     <dbl>
#>  1      350  0.000620
#>  2      350  0.000536
#>  3      351  0.000627
#>  4      351  0.000566
#>  5      352  0.000553
#>  6      352  0.000461
#>  7      353  0.000554
#>  8      353  0.000641
#>  9      354  0.000595
#> 10      354  0.000457
#> # ... with 1,202 more rows
#> --- Member: R ---
#> Object: source_spct [1,212 x 2]
#> Wavelength range 350 to 898.81 nm, step 0.36 to 0.47 nm 
#> Label: LedSaver_7.5W_R 
#> Measured on 2017-10-02 12:34:07 UTC 
#> Time unit 1s
#> 
#> # A tibble: 1,212 x 2
#>    w.length  s.e.irrad
#>  *    <dbl>      <dbl>
#>  1      350 0         
#>  2      350 0         
#>  3      351 0         
#>  4      351 0.000136  
#>  5      352 0.000134  
#>  6      352 0         
#>  7      353 0         
#>  8      353 0         
#>  9      354 0.00000319
#> 10      354 0         
#> # ... with 1,202 more rows
#> --- Member: G ---
#> Object: source_spct [1,212 x 2]
#> Wavelength range 350 to 898.81 nm, step 0.36 to 0.47 nm 
#> Label: LedSaver_7.5W_G 
#> Measured on 2017-10-02 12:36:20 UTC 
#> Time unit 1s
#> 
#> # A tibble: 1,212 x 2
#>    w.length s.e.irrad
#>  *    <dbl>     <dbl>
#>  1      350 0.0000134
#>  2      350 0.0000573
#>  3      351 0        
#>  4      351 0.0000655
#>  5      352 0.0000104
#>  6      352 0.0000623
#>  7      353 0        
#>  8      353 0.000270 
#>  9      354 0.0000693
#> 10      354 0        
#> # ... with 1,202 more rows
#> --- Member: B ---
#> Object: source_spct [1,212 x 2]
#> Wavelength range 350 to 898.81 nm, step 0.36 to 0.47 nm 
#> Label: LedSaver_7.5W_B 
#> Measured on 2017-10-02 12:38:13 UTC 
#> Time unit 1s
#> 
#> # A tibble: 1,212 x 2
#>    w.length s.e.irrad
#>  *    <dbl>     <dbl>
#>  1      350  0.000428
#>  2      350  0.000338
#>  3      351  0.000430
#>  4      351  0.000290
#>  5      352  0.000709
#>  6      352  0.000282
#>  7      353  0.000518
#>  8      353  0.000312
#>  9      354  0.000489
#> 10      354  0.000259
#> # ... with 1,202 more rows
#> 
#> --- END ---