Spectra acquired with LI-COR LI-1800

A collection of spectra measured with a LI-COR LI-1800 spectroradiometer. Member spectra containing the wavelengths at a 0.5 nm, 1.0 nm, or 2.0 nm interval and tabulated values of spectral irradiance for different lamps. As absolute values are not meaningful because measuring distances are in most cases unknown data have been normalized to one at the wavelength of maximum energy irradiance.

licor

Format

A vector of character strings.

Details

licor is a character vector with indexes to members of lamps.mspct incandescent.60w osram.36w.25 philips.pls11w.827 philips.tld36w.15 philips.tld36w.18 philips.tld36w.83 philips.tld36w.865 philips.tld36w.89 philips.tld36w.92 philips.tll36w.950 sylvania.215w.vho osram.HQIT.400w osram.super.vialox

The variables in each member spectrum are as follows:

w.length (nm)

s.e.irrad (relative energy based units)

Note

Instrument used: LI-COR LI1800 scanning single monochromator spectroradiometer with a cosine corrected input optics. Recently calibrated with a LI-COR calibration lamp. Measurements done by Pedro J. Aphalo between 1992 and 2001.

Examples

lamps.mspct[licor]

#> Object: source_mspct [13 x 1]
#> --- Member: incandescent.60w ---
#> Object: source_spct [601 x 2]
#> Wavelength range 300 to 900 nm, step 1 nm 
#> Label: File: Incandescent.60W.PRN 
#> Measured on 0000-08-18 15:27:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 900 nm 
#> 
#> # A tibble: 601 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300  0.000665
#>  2      301  0.000325
#>  3      302  0.000956
#>  4      303  0.000702
#>  5      304  0.000306
#>  6      305  0.000825
#>  7      306  0.00103 
#>  8      307  0.000361
#>  9      308  0.000497
#> 10      309  0.000627
#> # ... with 591 more rows
#> --- Member: osram.36w.25 ---
#> Object: source_spct [301 x 2]
#> Wavelength range 300 to 900 nm, step 2 nm 
#> Label: File: Osram.36W.25.PRN 
#> Measured on 0000-06-09 09:00:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 436 nm 
#> 
#> # A tibble: 301 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300 0.000299 
#>  2      302 0.000205 
#>  3      304 0.000354 
#>  4      306 0.0000757
#>  5      308 0.000292 
#>  6      310 0.000247 
#>  7      312 0.000239 
#>  8      314 0.000330 
#>  9      316 0.000320 
#> 10      318 0.000155 
#> # ... with 291 more rows
#> --- Member: osram.hqit.400w ---
#> Object: source_spct [301 x 2]
#> Wavelength range 300 to 900 nm, step 2 nm 
#> Label: File: Osram.HQIT.400W.PRN 
#> Measured on 0000-10-11 18:57:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 592 nm 
#> 
#> # A tibble: 301 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300  0.000382
#>  2      302  0.000641
#>  3      304  0.000615
#>  4      306  0.000761
#>  5      308  0.000795
#>  6      310  0.00102 
#>  7      312  0.00141 
#>  8      314  0.00162 
#>  9      316  0.00161 
#> 10      318  0.00170 
#> # ... with 291 more rows
#> --- Member: osram.super.vialox ---
#> Object: source_spct [301 x 2]
#> Wavelength range 300 to 900 nm, step 2 nm 
#> Label: File: Osram.Super.Vialox.PRN 
#> Measured on 0000-10-17 13:54:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 820 nm 
#> 
#> # A tibble: 301 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300  0.000235
#>  2      302  0.000196
#>  3      304  0.000237
#>  4      306  0.000378
#>  5      308  0.000553
#>  6      310  0.000902
#>  7      312  0.00139 
#>  8      314  0.00198 
#>  9      316  0.00270 
#> 10      318  0.00348 
#> # ... with 291 more rows
#> --- Member: philips.pls11w.827 ---
#> Object: source_spct [301 x 2]
#> Wavelength range 300 to 900 nm, step 2 nm 
#> Label: File: Philips.PLS11W.827.PRN 
#> Measured on 0000-10-18 10:31:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 612 nm 
#> 
#> # A tibble: 301 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300  0.000135
#>  2      302  0.000370
#>  3      304  0.000351
#>  4      306  0.000227
#>  5      308  0.00240 
#>  6      310  0.00804 
#>  7      312  0.0151  
#>  8      314  0.0171  
#>  9      316  0.0115  
#> 10      318  0.00536 
#> # ... with 291 more rows
#> --- Member: philips.tld36w.15 ---
#> Object: source_spct [601 x 2]
#> Wavelength range 300 to 900 nm, step 1 nm 
#> Label: File: Philips.TLD36W.15.PRN 
#> Measured on 0000-08-23 11:56:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 660 nm 
#> 
#> # A tibble: 601 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300  0.000196
#>  2      301  0.000959
#>  3      302  0.000752
#>  4      303  0.00110 
#>  5      304  0.00126 
#>  6      305  0.000885
#>  7      306  0.000521
#>  8      307  0.00153 
#>  9      308  0.000167
#> 10      309  0.000493
#> # ... with 591 more rows
#> --- Member: philips.tld36w.18 ---
#> Object: source_spct [601 x 2]
#> Wavelength range 300 to 900 nm, step 1 nm 
#> Label: File: Philips.TLD36W.18.PRN 
#> Measured on 0000-08-23 13:13:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 437 nm 
#> 
#> # A tibble: 601 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300 0.000293 
#>  2      301 0.000287 
#>  3      302 0.0000802
#>  4      303 0.0000785
#>  5      304 0.000192 
#>  6      305 0.000302 
#>  7      306 0.000148 
#>  8      307 0.000182 
#>  9      308 0.000250 
#> 10      309 0.000351 
#> # ... with 591 more rows
#> --- Member: philips.tld36w.83 ---
#> Object: source_spct [601 x 2]
#> Wavelength range 300 to 900 nm, step 1 nm 
#> Label: File: Philips.TLD36W.83.PRN 
#> Measured on 0000-08-23 13:31:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 613 nm 
#> 
#> # A tibble: 601 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300  0.000381
#>  2      301  0.000497
#>  3      302  0.000426
#>  4      303  0.000476
#>  5      304  0.000467
#>  6      305  0.000343
#>  7      306  0.000449
#>  8      307  0.000606
#>  9      308  0.00130 
#> 10      309  0.00229 
#> # ... with 591 more rows
#> --- Member: philips.tld36w.865 ---
#> Object: source_spct [601 x 2]
#> Wavelength range 300 to 900 nm, step 1 nm 
#> Label: File: Philips.TLD36W.865.PRN 
#> Measured on 0000-08-23 16:32:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 546 nm 
#> 
#> # A tibble: 601 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300  0.000369
#>  2      301  0.000812
#>  3      302  0.000530
#>  4      303  0.000779
#>  5      304  0.00110 
#>  6      305  0.000499
#>  7      306  0.000898
#>  8      307  0.00112 
#>  9      308  0.00181 
#> 10      309  0.00278 
#> # ... with 591 more rows
#> --- Member: philips.tld36w.89 ---
#> Object: source_spct [601 x 2]
#> Wavelength range 300 to 900 nm, step 1 nm 
#> Label: File: Philips.TLD36W.89.PRN 
#> Measured on 0000-08-23 16:59:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 613 nm 
#> 
#> # A tibble: 601 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300  0.000477
#>  2      301  0.000560
#>  3      302  0.00100 
#>  4      303  0.000626
#>  5      304  0.000965
#>  6      305  0.000860
#>  7      306  0.000760
#>  8      307  0.000746
#>  9      308  0.00114 
#> 10      309  0.00176 
#> # ... with 591 more rows
#> ..........................
#> 3 other member spectra not shown
#> 
#> --- END ---

lamps.mspct[["incandescent.60w"]]

#> Object: source_spct [601 x 2]
#> Wavelength range 300 to 900 nm, step 1 nm 
#> Label: File: Incandescent.60W.PRN 
#> Measured on 0000-08-18 15:27:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 900 nm 
#> 
#> # A tibble: 601 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300  0.000665
#>  2      301  0.000325
#>  3      302  0.000956
#>  4      303  0.000702
#>  5      304  0.000306
#>  6      305  0.000825
#>  7      306  0.00103 
#>  8      307  0.000361
#>  9      308  0.000497
#> 10      309  0.000627
#> # ... with 591 more rows

lamps.mspct$incandescent.60w

#> Object: source_spct [601 x 2]
#> Wavelength range 300 to 900 nm, step 1 nm 
#> Label: File: Incandescent.60W.PRN 
#> Measured on 0000-08-18 15:27:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 900 nm 
#> 
#> # A tibble: 601 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300  0.000665
#>  2      301  0.000325
#>  3      302  0.000956
#>  4      303  0.000702
#>  5      304  0.000306
#>  6      305  0.000825
#>  7      306  0.00103 
#>  8      307  0.000361
#>  9      308  0.000497
#> 10      309  0.000627
#> # ... with 591 more rows