Spectra acquired with LI-COR LI-1800

Lamp emission spectra measured with a LI-COR LI-1800 spectroradiometer. Member spectra containing the wavelengths at a 0.5 nm, 1.0 nm, or 2.0 nm interval and tabulated values of spectral irradiance for different lamps. As absolute values are not meaningful because measuring distances are in most cases unknown data have been normalized to one at the wavelength of maximum energy irradiance.

licor_lamps

Format

A vector of character strings.

Details

licor is a character vector with indexes to members of lamps.mspct

Note

Instrument used: LI-COR LI1800 scanning single monochromator spectroradiometer with a cosine corrected input optics. Recently calibrated with a LI-COR calibration lamp. Measurements done by Pedro J. Aphalo between 1992 and 2001.

Examples

licor_lamps

#>  [1] "incandescent.60w"   "osram.36w.25"       "osram.hqit.400w"   
#>  [4] "osram.super.vialox" "philips.pls11w.827" "philips.tld36w.15" 
#>  [7] "philips.tld36w.18"  "philips.tld36w.83"  "philips.tld36w.865"
#> [10] "philips.tld36w.89"  "philips.tld36w.92"  "philips.tll36w.950"
#> [13] "sylvania.215w.vho"