Spectra acquired with Macam SR-9010-PC

A collection of lamp emission spectra measured with a Macam SR-9010-PC spectroradiometer.

macam

Format

A vector of character strings.

Details

macam is a character vector with indexes to members of lamps.mspct with spectral emission data for Q-Panel UVB-313 and UVA-340 40W fluorescent tubes, and for Philips TL12 40W fluorescent tubes.

The variables in each member spectrum are as follows:

w.length (nm)

s.e.irrad (relative energy based units)

Note

Instrument used: Macam SR-9010-PC scanning double monochromator spectroradiometer with a cosine corrected input optics. Recently calibrated. The lamps were probably driven by a high frequency electronic ballast. Measured between 1996 to 1998.

References

Aphalo, P J, R Tegelberg, and R Julkunen-Tiitto. 1999. The Modulated UV-B Irradiation System at the University of Joensuu.” Biotronics 28: 109–120. http://ci.nii.ac.jp/naid/110006175827/en.

Examples

lamps.mspct[macam]

#> Object: source_mspct [3 x 1]
#> --- Member: qpanel.uvb313 ---
#> Object: source_spct [561 x 2]
#> Wavelength range 240 to 800 nm, step 1 nm 
#> Time unit 1s
#> Spectral data normalized to 1 at 435 nm 
#> 
#> # A tibble: 561 x 2
#>    w.length s.e.irrad
#>       <int>     <dbl>
#>  1      240         0
#>  2      241         0
#>  3      242         0
#>  4      243         0
#>  5      244         0
#>  6      245         0
#>  7      246         0
#>  8      247         0
#>  9      248         0
#> 10      249         0
#> # ... with 551 more rows
#> --- Member: qpanel.uva340 ---
#> Object: source_spct [561 x 2]
#> Wavelength range 240 to 800 nm, step 1 nm 
#> Time unit 1s
#> Spectral data normalized to 1 at 435 nm 
#> 
#> # A tibble: 561 x 2
#>    w.length s.e.irrad
#>       <int>     <dbl>
#>  1      240         0
#>  2      241         0
#>  3      242         0
#>  4      243         0
#>  5      244         0
#>  6      245         0
#>  7      246         0
#>  8      247         0
#>  9      248         0
#> 10      249         0
#> # ... with 551 more rows
#> --- Member: philips.tl12.mc ---
#> Object: source_spct [561 x 2]
#> Wavelength range 240 to 800 nm, step 1 nm 
#> Time unit 1s
#> Spectral data normalized to 1 at 435 nm 
#> 
#> # A tibble: 561 x 2
#>    w.length s.e.irrad
#>       <int>     <dbl>
#>  1      240         0
#>  2      241         0
#>  3      242         0
#>  4      243         0
#>  5      244         0
#>  6      245         0
#>  7      246         0
#>  8      247         0
#>  9      248         0
#> 10      249         0
#> # ... with 551 more rows
#> 
#> --- END ---

lamps.mspct[["qpanel.uvb313"]]

#> Object: source_spct [561 x 2]
#> Wavelength range 240 to 800 nm, step 1 nm 
#> Time unit 1s
#> Spectral data normalized to 1 at 435 nm 
#> 
#> # A tibble: 561 x 2
#>    w.length s.e.irrad
#>       <int>     <dbl>
#>  1      240         0
#>  2      241         0
#>  3      242         0
#>  4      243         0
#>  5      244         0
#>  6      245         0
#>  7      246         0
#>  8      247         0
#>  9      248         0
#> 10      249         0
#> # ... with 551 more rows

lamps.mspct$qpanel.uvb313

#> Object: source_spct [561 x 2]
#> Wavelength range 240 to 800 nm, step 1 nm 
#> Time unit 1s
#> Spectral data normalized to 1 at 435 nm 
#> 
#> # A tibble: 561 x 2
#>    w.length s.e.irrad
#>       <int>     <dbl>
#>  1      240         0
#>  2      241         0
#>  3      242         0
#>  4      243         0
#>  5      244         0
#>  6      245         0
#>  7      246         0
#>  8      247         0
#>  9      248         0
#> 10      249         0
#> # ... with 551 more rows