Spectra acquired with Ocean Optics Maya2000 Pro

Spectra measured with a Ocean Optics Maya2000 Pro spectroradiometer. Datasets contain wavelengths at variable intervals and corresponding values for spectral emittance from different lamps. Absolute values are not meaningful because data have been normalized to one at the wavelength of maximum energy irradiance.

oo_maya_lamps

Format

A vector of character strings.

Details

oo_maya_lamps is a character vector with indexes to members of lamps.mspct

Note

Instrument used: Ocean Optics Maya2000 Pro single-monochromator array spectroradiometer with a Bentham cosine corrected input optics. A complex set of corrections and calibration procedure used. Raw spectral data processed with R package MayaCalc. The source_spct object contains a comment with additional information on the measurement and data processing. Measurements done by Pedro J. Aphalo.

Examples

oo_maya_lamps

#>  [1] "airam.cf.15w.2700k"              "airam.led.11w.4000k"            
#>  [3] "airam.led.oiva.9w.3000k"         "airam.spiraali"                 
#>  [5] "amaran.al.m9"                    "germicidal"                     
#>  [7] "osram.classic.20w"               "osram.conc.spot.60w"            
#>  [9] "osram.l36w.840"                  "osram.led.10w.2700k.classicstar"
#> [11] "osram.led.8w.2700k"              "philips.tl5.35w.830he"          
#> [13] "philips.tld.36w.18"              "pirkka.halogen.53w"             
#> [15] "toshiba.led.9.5w.2700k"          "v.light.led.2w.6000k"           
#> [17] "valoya.b100.ap67"