photobiologyLamps: Spectral Irradiance Data for Lamps

Spectral emission data for some frequently used lamps excluding led emitting diodes (LEDs) available as electronic components. Original spectral irradiance data for incandescent-, LED- and discharge lamps are included. They are complemented by data on the effect of temperature on the emission by fluorescent tubes.

Details

This package contains emission spectra for different types of lamps including LED lamps sold as fully assembled light bulbs or luminaires but excluding LEDs sold as components requiring separate electronic driver modules or circuits. Lamps have been measured with an assortment of different instruments over more than 25 years in our lab or by collaborators. Some of the lamp types are no longer in production but the data are relevant for the interpretation of old scientific publications. Data for lamps emitting in the UV and VIS regions of the spectrum are represented in these data sets. In addition a set of spectral emission for UVB lamps exposed to different ambient temperatures is include.

The package contains two collections of spectra for different lamps all of them measured at air temperatures between 20 C and 25 C and a series of vectors to be used as indexes to extract different subsets of spectra. These spectral data are normalized to spectral energy irradiance equal one at their maximum (strongest emission peak).

Temperature response data is included as a separate collection of spectra both as a source_mspct object and as a source_spct object.

A third collection of spectra contains data for the 16 different _colors_ of light emitted by a household four channel LED bulb.

Warning!

None of the spectral data included in this package are based on supplier's specifications and are only for information. The exact emission spectrum depends to some extent on testing conditions, but more importantly among individual lamps. It is not uncommon for suppliers to revise specifications during the lifetime of a product type without changing the product denomination. There can be also some variation from batch to batch. It is also important to make sure that the type considered exactly matches. Frequently variations of a lamp type, with type denomination differing in a single letter or number, have quite different spectral characteristics. In other words, the data provided here are not a substitute for actual measurements of radiation emission and spectrum of the lamps actually used in a given piece of scientific research. For less demanding situations the data are in most cases reliable enough but perfect agreement with measurements on other lamps of the same exact type should not be expected. This is specially true for those data measured more than a few years ago when compared to lamps being currently sold under the same type denomination.

See also

Packages photobiology-package and ggspectra-package.

Examples

library(photobiologyLamps)
library(photobiologyWavebands)

#> For news about 'photobiologyWavebands', please, see http://www.r4photobiology.info/

#> For on-line documentation see http://docs.r4photobiology.info/photobiologyWavebands/

# extract one spectrum
lamps.mspct$incandescent.60w

#> Object: source_spct [601 x 2]
#> Wavelength range 300 to 900 nm, step 1 nm 
#> Label: File: Incandescent.60W.PRN 
#> Measured on 0000-08-18 15:27:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 900 nm 
#> 
#> # A tibble: 601 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300  0.000665
#>  2      301  0.000325
#>  3      302  0.000956
#>  4      303  0.000702
#>  5      304  0.000306
#>  6      305  0.000825
#>  7      306  0.00103 
#>  8      307  0.000361
#>  9      308  0.000497
#> 10      309  0.000627
#> # ... with 591 more rows

lamps.mspct[["incandescent.60w"]]

#> Object: source_spct [601 x 2]
#> Wavelength range 300 to 900 nm, step 1 nm 
#> Label: File: Incandescent.60W.PRN 
#> Measured on 0000-08-18 15:27:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 900 nm 
#> 
#> # A tibble: 601 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300  0.000665
#>  2      301  0.000325
#>  3      302  0.000956
#>  4      303  0.000702
#>  5      304  0.000306
#>  6      305  0.000825
#>  7      306  0.00103 
#>  8      307  0.000361
#>  9      308  0.000497
#> 10      309  0.000627
#> # ... with 591 more rows

# using one spectrum in a calculation
q_ratio(lamps.mspct$incandescent.60w, Blue(), Green())

#>  Blue.ISO: Green.ISO(q:q) 
#>                 0.2734537 
#> attr(,"radiation.unit")
#> [1] "q:q ratio"

# extracting all the spectra measured with a given instrument
lamps.mspct[bentham_lamps]

#> Object: source_mspct [2 x 1]
#> --- Member: philips.tl01 ---
#> Object: source_spct [301 x 2]
#> Wavelength range 250 to 400 nm, step 0.5 nm 
#> Time unit 1s
#> Spectral data normalized to 1 at 311.5 nm 
#> 
#> # A tibble: 301 x 2
#>    w.length    s.e.irrad
#>       <dbl>        <dbl>
#>  1      250 0.000000249 
#>  2      250 0.0000000624
#>  3      251 0.0000000340
#>  4      252 0.0000000166
#>  5      252 0.0000000261
#>  6      252 0.000000468 
#>  7      253 0.0000280   
#>  8      254 0.0000979   
#>  9      254 0.0000102   
#> 10      254 0.000000201 
#> # ... with 291 more rows
#> --- Member: philips.tl12 ---
#> Object: source_spct [301 x 2]
#> Wavelength range 250 to 400 nm, step 0.5 nm 
#> Time unit 1s
#> Spectral data normalized to 1 at 313 nm 
#> 
#> # A tibble: 301 x 2
#>    w.length   s.e.irrad
#>       <dbl>       <dbl>
#>  1      250 0.000000683
#>  2      250 0          
#>  3      251 0          
#>  4      252 0          
#>  5      252 0.000000365
#>  6      252 0.00000135 
#>  7      253 0.0000379  
#>  8      254 0.000153   
#>  9      254 0.0000109  
#> 10      254 0.000000978
#> # ... with 291 more rows
#> 
#> --- END ---