Spectral irradiance of UVB lamps at different temperatures.

Dataset of spectral irradiance from a set of two Q-Panel UVB-313 40W fluorescent tubes, measured at different temperatures.

qp_uvb313_temp.spct

qp_uvb313_temp.mspct

Format

qp_uvb313_temp.spct is a "source_spct" object with 777 rows (290 nm to 400 nm, 1.0 nm step) and 3 variables.

The variables are as follows:

temperature (C)

w.length (nm)

s.e.irrad (W m-2 nm-1)

qp_uvb313_temp.mspct is a "source_mspct" object containing a collection of "source_spct" objects, each with 111 rows (290 nm to 400 nm, 1.0 nm step) and 2 variables.

The variables in the member spectra are as follows:

w.length (nm)

s.e.irrad (W m-2 nm-1)

Details

Absolute values are comparable among the different temperatures. The lamps and the cosine diffuser were located inside a Conviron growth chamber. The measurements were done with a recently calibrated Macam double-monochromator spectroradiometer located outside the chamber and maintained at constant room temperature of 22 C. After each change in the temperature controller set-point enough time was allowed after the temperature had stabilized inside the chamber, for the output of the lamps to also become stable. qp_uvb313_temp.spct contains data for different temperatures in a "tidy" long form with a factor indicating the temperature at which the lamp was during measurement.

qp_uvb313_temp.mspct contains the same data as a collection of spectra stored in a "source_mspct" object, with member "source_spct" objects named minus5C, plus00C, plus05C, plus10C, plus20C, plus30C, and plus35C.

Note

Instrument used: Macam SR-9010-PC scanning double monochromator spectroradiometer with a cosine corrected input optics. Recently calibrated. The lamps were driven by a high frequency electronic ballast.

References

Aphalo, P J, R Tegelberg, and R Julkunen-Tiitto. 1999. The Modulated UV-B Irradiation System at the University of Joensuu.” Biotronics 28: 109–120. http://ci.nii.ac.jp/naid/110006175827/en.

Examples

qp_uvb313_temp.mspct[["plus20C"]]

#> Object: source_spct [111 x 3]
#> containing 7 spectra in long form
#> Wavelength range 290 to 400 nm, step 1 nm 
#> Time unit 1s
#> 
#> # A tibble: 111 x 3
#>    temperature w.length s.e.irrad
#>  *       <dbl>    <int>     <dbl>
#>  1         20.      290 0.       
#>  2         20.      291 0.       
#>  3         20.      292 0.0000496
#>  4         20.      293 0.000379 
#>  5         20.      294 0.00126  
#>  6         20.      295 0.00321  
#>  7         20.      296 0.00673  
#>  8         20.      297 0.0182   
#>  9         20.      298 0.0226   
#> 10         20.      299 0.0289   
#> # ... with 101 more rows