Spectral data for Lamps of different colors

Names of datasets containing the wavelengths and tabulated values spectral emittance for the lamps of different types from various suppliers.

uv_lamps

white_lamps

blue_lamps

amber_lamps

red_lamps

Format

A vector of character strings.

See also

lamps.mspct

Examples

uv_lamps

#> [1] "germicidal"      "philips.tl01"    "philips.tl12"    "philips.tl12.mc"
#> [5] "qpanel.uva340"   "qpanel.uvb313"  

blue_lamps

#> [1] "philips.tld36w.18"  "philips.tld.36w.18"

red_lamps

#> [1] "philips.tld36w.15"

amber_lamps

#> [1] "osram.super.vialox"

white_lamps

#>  [1] "airam.cf.15w.2700k"              "airam.led.oiva.9w.3000k"        
#>  [3] "airam.spiraali"                  "incandescent.60w"               
#>  [5] "osram.36w.25"                    "osram.classic.20w"              
#>  [7] "osram.conc.spot.60w"             "osram.hqit.400w"                
#>  [9] "osram.l36w.840"                  "osram.led.10w.2700k.classicstar"
#> [11] "osram.led.8w.2700k"              "philips.pls11w.827"             
#> [13] "philips.tl5.35w.830he"           "philips.tld36w.83"              
#> [15] "philips.tld36w.865"              "philips.tld36w.89"              
#> [17] "philips.tld36w.92"               "philips.tll36w.950"             
#> [19] "pirkka.halogen.53w"              "sylvania.215w.vho"              
#> [21] "toshiba.led.9.5w.2700k"          "valoya.b100.ap67"               

# select lamps emitting in the amber, yellow, orange region
lamps.mspct[red_lamps]

#> Object: source_mspct [1 x 1]
#> --- Member: philips.tld36w.15 ---
#> Object: source_spct [601 x 2]
#> Wavelength range 300 to 900 nm, step 1 nm 
#> Label: File: Philips.TLD36W.15.PRN 
#> Measured on 0000-08-23 11:56:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 660 nm 
#> 
#> # A tibble: 601 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300  0.000196
#>  2      301  0.000959
#>  3      302  0.000752
#>  4      303  0.00110 
#>  5      304  0.00126 
#>  6      305  0.000885
#>  7      306  0.000521
#>  8      307  0.00153 
#>  9      308  0.000167
#> 10      309  0.000493
#> # ... with 591 more rows
#> 
#> --- END ---