Spectral irradiance for diverse lamps

A collection of lamp emission spectra measured with different spectroradiometers.

lamps.mspct

Format

A "source_mspct" object containing several "source_spct".

Details

The "source_mspct" object contains "source_spct" objects with spectral emission data.

The variables in each member spectrum are as follows:

w.length (nm)

s.e.irrad (relative energy based units)

Note

Please see the help corresponding to each instrument for details.

See also

bentham_lamps, licor_lamps, macam_lamps, oo_maya_lamps for information on the spectroradiometers, and photobiologyLEDs for spectra for LEDs available as electronic components.

Examples

uv_lamps

#> [1] "germicidal"      "philips.tl01"    "philips.tl12"    "philips.tl12.mc"
#> [5] "qpanel.uva340"   "qpanel.uvb313"  

union(qpanel_lamps, uv_lamps)

#> [1] "qpanel.uva340"   "qpanel.uvb313"   "germicidal"      "philips.tl01"   
#> [5] "philips.tl12"    "philips.tl12.mc"

lamps.mspct[union(philips_lamps, red_lamps)]

#> Object: source_mspct [11 x 1]
#> --- Member: philips.pls11w.827 ---
#> Object: source_spct [301 x 2]
#> Wavelength range 300 to 900 nm, step 2 nm 
#> Label: File: Philips.PLS11W.827.PRN 
#> Measured on 0000-10-18 10:31:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 612 nm 
#> 
#> # A tibble: 301 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300  0.000135
#>  2      302  0.000370
#>  3      304  0.000351
#>  4      306  0.000227
#>  5      308  0.00240 
#>  6      310  0.00804 
#>  7      312  0.0151  
#>  8      314  0.0171  
#>  9      316  0.0115  
#> 10      318  0.00536 
#> # ... with 291 more rows
#> --- Member: philips.tld36w.83 ---
#> Object: source_spct [601 x 2]
#> Wavelength range 300 to 900 nm, step 1 nm 
#> Label: File: Philips.TLD36W.83.PRN 
#> Measured on 0000-08-23 13:31:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 613 nm 
#> 
#> # A tibble: 601 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300  0.000381
#>  2      301  0.000497
#>  3      302  0.000426
#>  4      303  0.000476
#>  5      304  0.000467
#>  6      305  0.000343
#>  7      306  0.000449
#>  8      307  0.000606
#>  9      308  0.00130 
#> 10      309  0.00229 
#> # ... with 591 more rows
#> --- Member: philips.tld36w.865 ---
#> Object: source_spct [601 x 2]
#> Wavelength range 300 to 900 nm, step 1 nm 
#> Label: File: Philips.TLD36W.865.PRN 
#> Measured on 0000-08-23 16:32:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 546 nm 
#> 
#> # A tibble: 601 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300  0.000369
#>  2      301  0.000812
#>  3      302  0.000530
#>  4      303  0.000779
#>  5      304  0.00110 
#>  6      305  0.000499
#>  7      306  0.000898
#>  8      307  0.00112 
#>  9      308  0.00181 
#> 10      309  0.00278 
#> # ... with 591 more rows
#> --- Member: philips.tld36w.89 ---
#> Object: source_spct [601 x 2]
#> Wavelength range 300 to 900 nm, step 1 nm 
#> Label: File: Philips.TLD36W.89.PRN 
#> Measured on 0000-08-23 16:59:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 613 nm 
#> 
#> # A tibble: 601 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300  0.000477
#>  2      301  0.000560
#>  3      302  0.00100 
#>  4      303  0.000626
#>  5      304  0.000965
#>  6      305  0.000860
#>  7      306  0.000760
#>  8      307  0.000746
#>  9      308  0.00114 
#> 10      309  0.00176 
#> # ... with 591 more rows
#> --- Member: philips.tld36w.92 ---
#> Object: source_spct [601 x 2]
#> Wavelength range 300 to 900 nm, step 1 nm 
#> Label: File: Philips.TLD36W.92.PRN 
#> Measured on 0000-08-23 16:41:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 547 nm 
#> 
#> # A tibble: 601 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300   0.00192
#>  2      301   0.00251
#>  3      302   0.00307
#>  4      303   0.00241
#>  5      304   0.00265
#>  6      305   0.00318
#>  7      306   0.00312
#>  8      307   0.00306
#>  9      308   0.00492
#> 10      309   0.00779
#> # ... with 591 more rows
#> --- Member: philips.tll36w.950 ---
#> Object: source_spct [601 x 2]
#> Wavelength range 300 to 900 nm, step 1 nm 
#> Label: File: Philips.TLL36W.950.PRN 
#> Measured on 0000-08-23 10:24:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 613 nm 
#> 
#> # A tibble: 601 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300 0.0000697
#>  2      301 0.0000682
#>  3      302 0.0000382
#>  4      303 0.0000561
#>  5      304 0.0000733
#>  6      305 0.0000359
#>  7      306 0.0000353
#>  8      307 0.0000346
#>  9      308 0.000136 
#> 10      309 0.000250 
#> # ... with 591 more rows
#> --- Member: philips.tl01 ---
#> Object: source_spct [301 x 2]
#> Wavelength range 250 to 400 nm, step 0.5 nm 
#> Time unit 1s
#> Spectral data normalized to 1 at 311.5 nm 
#> 
#> # A tibble: 301 x 2
#>    w.length    s.e.irrad
#>       <dbl>        <dbl>
#>  1      250 0.000000249 
#>  2      250 0.0000000624
#>  3      251 0.0000000340
#>  4      252 0.0000000166
#>  5      252 0.0000000261
#>  6      252 0.000000468 
#>  7      253 0.0000280   
#>  8      254 0.0000979   
#>  9      254 0.0000102   
#> 10      254 0.000000201 
#> # ... with 291 more rows
#> --- Member: philips.tl12 ---
#> Object: source_spct [301 x 2]
#> Wavelength range 250 to 400 nm, step 0.5 nm 
#> Time unit 1s
#> Spectral data normalized to 1 at 313 nm 
#> 
#> # A tibble: 301 x 2
#>    w.length   s.e.irrad
#>       <dbl>       <dbl>
#>  1      250 0.000000683
#>  2      250 0          
#>  3      251 0          
#>  4      252 0          
#>  5      252 0.000000365
#>  6      252 0.00000135 
#>  7      253 0.0000379  
#>  8      254 0.000153   
#>  9      254 0.0000109  
#> 10      254 0.000000978
#> # ... with 291 more rows
#> --- Member: philips.tl12.mc ---
#> Object: source_spct [561 x 2]
#> Wavelength range 240 to 800 nm, step 1 nm 
#> Time unit 1s
#> Spectral data normalized to 1 at 435 nm 
#> 
#> # A tibble: 561 x 2
#>    w.length s.e.irrad
#>       <int>     <dbl>
#>  1      240         0
#>  2      241         0
#>  3      242         0
#>  4      243         0
#>  5      244         0
#>  6      245         0
#>  7      246         0
#>  8      247         0
#>  9      248         0
#> 10      249         0
#> # ... with 551 more rows
#> --- Member: philips.tld36w.15 ---
#> Object: source_spct [601 x 2]
#> Wavelength range 300 to 900 nm, step 1 nm 
#> Label: File: Philips.TLD36W.15.PRN 
#> Measured on 0000-08-23 11:56:00 UTC 
#> Time unit 1s
#> Spectral data normalized to 1 at 660 nm 
#> 
#> # A tibble: 601 x 2
#>    w.length s.e.irrad
#>       <dbl>     <dbl>
#>  1      300  0.000196
#>  2      301  0.000959
#>  3      302  0.000752
#>  4      303  0.00110 
#>  5      304  0.00126 
#>  6      305  0.000885
#>  7      306  0.000521
#>  8      307  0.00153 
#>  9      308  0.000167
#> 10      309  0.000493
#> # ... with 591 more rows
#> ..........................
#> 1 other member spectra not shown
#> 
#> --- END ---

lamps.mspct$qpanel.uvb313

#> Object: source_spct [561 x 2]
#> Wavelength range 240 to 800 nm, step 1 nm 
#> Time unit 1s
#> Spectral data normalized to 1 at 435 nm 
#> 
#> # A tibble: 561 x 2
#>    w.length s.e.irrad
#>       <int>     <dbl>
#>  1      240         0
#>  2      241         0
#>  3      242         0
#>  4      243         0
#>  5      244         0
#>  6      245         0
#>  7      246         0
#>  8      247         0
#>  9      248         0
#> 10      249         0
#> # ... with 551 more rows